requirements.txt
playwright==1.40.0
requests==2.31.0
openai==1.35.0
instructor==1.3.0
pydantic==2.7.0
tqdm==4.66.0
yfinance==0.2.28
pandas>=2.0.0
python-dotenv==1.0.0